+++ R/MizerParams-class.R
@@ -496,14 +496,14 @@ emptyParams <- function(species_params,
                 is.data.frame(gear_params),
                 no_w > 10)
     
-    given_species_params <- validSpeciesParams(species_params)
+    given_species_params <- validGivenSpeciesParams(species_params)
     
     ## Set defaults ----
     if (is.na(min_w_pp)) min_w_pp <- 1e-12
     species_params <- set_species_param_default(species_params, "w_min", min_w)
     min_w <- min(species_params$w_min)
     
-    species_params <- completeSpeciesParams(species_params)
+    species_params <- validSpeciesParams(species_params)
     gear_params <- validGearParams(gear_params, species_params)
     
     if (is.na(max_w)) {
@@ -814,7 +814,13 @@ dw_full <- function(params) {
     params@dw_full
 }
 
-#' Validate MizerParams object and upgrade if necessary using [upgradeParams()].
+#' Validate MizerParams object and upgrade if necessary
+#' 
+#' Checks that the given MizerParams object is valid and upgrades it if
+#' necessary by calling [upgradeParams()].
+#' 
+#' Besides upgrading if necessary, the only change that may be made to the
+#' given MizerParams object is that the `w_min_idx` slot is recalculated.
 #' 
 #' @param params The MizerParams object to validate
 #' @return A valid MizerParams object